Change dict to bidict
In this commit, the key dictionary of encoder objects, _parent_label_to_islot to bidicts. This is because there is a bijective map between parent nodes and the islots they send their tensors to.

Other than that, Concat and Clone Nodes are removed from basic since they can be implemented directly as part of InnerNodes.